feat(backend): add validation and common testimonial files
