--- apps/backend/src/api/admin/custom/route.ts
@@ -1,8 +1,0 @@
-import { MedusaRequest, MedusaResponse } from '@medusajs/medusa';
-
-export async function GET(
-  req: MedusaRequest,
-  res: MedusaResponse,
-): Promise<void> {
-  res.sendStatus(200);
-}
