+++ apps/backend/src/api/admin/testimonials/route.ts
@@ -1,0 +1,6 @@
+import {
+  testimonialsGET,
+  testimonialsPOST,
+} from '../../../common/testimonials/routes';
+
+export { testimonialsGET as GET, testimonialsPOST as POST };
